Feature/khady2

+++ routes/api.php
@@ -88,22 +88,25 @@ Route::middleware(['is_jardinier'])->group(function () {
     });
 });
 
-Route::middleware(['is_jardinier'])->group(function (){
+
     Route::controller(CategoriesController::class)->group(function (){
-        Route::post('ModifierCategorie/{id}', 'update');
-        Route::post('AjouterCategorie', 'store');
-        Route::delete('SupprimerCategorie/{categories}', 'destroy');
-        Route::post('PublierVideo','publiervideo');
+        Route::post('ModifierCategorie/{id}', 'update')->middleware('checkadmin');
+        Route::post('AjouterCategorie', 'store')->middleware('checkadmin');
+        Route::delete('SupprimerCategorie/{categories}', 'destroy')->middleware('checkadmin');
+        Route::post('PublierVideo','publiervideo')->middleware('is_jardinier');
         Route::get('recupererVideo','index');
         Route::get('DetailVideo/{id}','show');
-        Route::post('RemplacerVideo/{id}','modifier');
-        Route::delete('supprimerVideo/{id}','effacer');
+        Route::post('RemplacerVideo/{id}','modifier')->middleware('is_jardinier');
+        Route::delete('supprimerVideo/{id}','effacer')->middleware(['is_jardinier','checkadmin']);
+        Route::get('listCategorie','listCategorie')->middleware(['is_jardinier','checkadmin']);
     });
-});
+
 
 Route::controller(MessageriesController::class)->group(function (){
     Route::post('EnvoyerMessage/{id}', 'sendMessage');
     Route::post('RepondreMessage/{message_id}', 'repondreMessage');
-    Route::get('RecupererMessage/{message_id}', 'getMessages');
+    Route::get('RecupererMessage/{envoyeurId}', 'getMessages');
+    Route::post('modifierMessage/{messageId}', 'modifierMessage');
+    Route::delete('SupprimerMessage/{messageId}', 'supprimerMessage');
 
 });
